Brain Tumor Analysis/requirements.txt
transformers~=4.44.0
torch~=2.4.0
datasets~=2.20.0
timm~=1.0.8
matplotlib~=3.8.1
numpy~=1.26.2
Pillow~=10.1.0
pandas~=2.1.3